requirements.txt
pandas
seaborn
matplotlib
numpy
requests
python-dotenv
